refactor: move function to utils and import
